Move anonymization to separate file
