+++ worker/src/handlers/statement/statementForwarding/statementForwardingRequestHandler.js
@@ -1,5 +1,5 @@
 import { post } from 'axios';
-import { assign, isPlainObject, get, set } from 'lodash';
+import { assign, isPlainObject } from 'lodash';
 import { PassThrough } from 'stream';
 import highland from 'highland';
 import getAttachments from '@learninglocker/xapi-statements/dist/service/utils/getAttachments';
@@ -16,59 +16,9 @@ import {
 } from 'lib/constants/statements';
 import * as Queue from 'lib/services/queue';
 import getStatementsRepo from './getStatementsRepo';
-const crypto = require('crypto');
+import pseudonymizeXAPIStatement from './statementAnonymization.js';
 
 const objectId = mongoose.Types.ObjectId;
-const generatePseudonym = (personalInformation) => {
-  // Hash the personal information using the chosen hash function
-  const hash = crypto.createHash(hashFunction);
-  hash.update(personalInformation);
-  let result =  hash.digest('hex');
-  if ((typeof personalInformation == "string") && !!(personalInformation.match("mailto"))) {
-    result = "mailto:" + result + "@anonymous.org";
-  }
-
-  return result;
-  
-}
-
-const pseudonymizeXAPIStatement = (xAPIStatement) => {
-  const fieldsToAnonymize = [
-    'statement.actor.mbox', 
-    'statement.actor.mbox_sha1sum', 
-    'statement.actor.account.email', 
-    'statement.actor.account.name', 
-    'statement.actor.account.homePage', 
-    'statement.actor.openid', 
-    'statement.actor.name', 
-    'statement.actor.member', 
-    'agents', 
-    'relatedAgents', 
-    'registrations',
-    'statement.context.registration'];
-  fieldsToAnonymize.forEach(field => {
-    try {
-    const personalInformation = get(xAPIStatement, field);
-    if (personalInformation) {
-      if (personalInformation && Array.isArray(personalInformation)) {
-        let pseudonym =  personalInformation.map(pi => {
-          return generatePseudonym(pi);
-        });
-        set(xAPIStatement, field, pseudonym);
-      } else {
-        let pseudonym = generatePseudonym(personalInformation);
-        set(xAPIStatement, field, pseudonym);
-      }
-      
-    }
-  } catch(ep){console.error(ep)}
-  });
-  return xAPIStatement;
-}
-
-
-// Choose a hash function
-const hashFunction = 'sha256';
 
 const generateHeaders = (statementForwarding, statement) => {
   const statementForwardingModel = new StatementForwarding(statementForwarding);
